`Empty` Element
The empty element, contrary to the `Marker`, is a truly
empty, thin element that does nothing. It can be used
for bookkeeping, e.g., as the empty result of a search
or as an indicator for an invalid element.

For all other purposes, use `Marker`.

Renamed from `Null` because `Null`/`null`/`None`/`none`
are reserved keywords in the file formats and programming
languages we support in PALS.